% Generated by roxygen2 (4.0.1): do not edit by hand
\name{category2pathway}
\alias{category2pathway}
\title{Condense COG/KEGG counts to metabolic pathways.}
\usage{
category2pathway(wgs_gene_count, pathway.file)
}
\arguments{
\item{wgs_gene_count}{data frame with COG/KEGG counts}

\item{pathway.file}{array with pathways composed of COG/KEGG annotations}
}
\description{
This function will collapses COG/KEGG counts into metabolic pathways.
}
\examples{
# Simulated data
  head(df.wgs.counts)
  cogpathway=path.file
  df.wgs.pathway.counts=category2pathway(df.wgs.counts,cogpathway)

# Mouse data
  head(mouse.cog)
  cogpathway=category2pathway.cog
  mouse.cog=category2pathway(mouse.cog,cogpathway)
}
\keyword{pathway}

