\name{dissim_pair_compare}
\alias{dissim_pair_compare}
\title{
[text]}
\description{
[text]}
\usage{
dissim_pair_compare(otu.normed, meta, grouping_variable, group_a, group_b)
}
\arguments{
  \item{otu.normed}{
Normatlized dataframe with relative abundance of OTUs}
  \item{meta}{
metadata}
  \item{grouping_variable}{
Ex. "bmi_group""}
  \item{group_a}{
Ex. "NW"}  
\item{group_b}{
Ex. "OB"}
}



\details{
[text]}
\value{
[text]
}
\references{
[text]}
\author{
[text]}
\note{
[text]}


\seealso{
[text]}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}

\keyword{ [text] }

