\name{dissim_pair_graph}
\alias{dissim_pair_graph}
\title{
[text]}
\description{
[text]}
\usage{
dissim_pair_graph(d.index)
}
\arguments{
  \item{d.index}{
the reult of dissim_compare}

}



\details{
[text]}
\value{
[text]
}
\references{
[text]}
\author{
[text]}
\note{
[text]}


\seealso{
[text]}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}

\keyword{ [text] }

