% Generated by roxygen2 (4.0.1): do not edit by hand
\name{diversity_table_compare}
\alias{diversity_table_compare}
\title{Test for group differences in alpha diversity.}
\usage{
diversity_table_compare(meta_group, div_table)
}
\arguments{
\item{div_table}{data frame of alpha diversity measures for each sample}

\item{meta_group}{vector that contains sample groups}
}
\description{
This function tests for group differences in alpha diversity and returns a table
with alpha diversity mean/sd for all samples as well as group mean/sd.
}
\examples{
# Test
diversity_table_compare(table)
test=diversity_table_compare(meta.grp.sort$bmi_group,table); test
}
\keyword{alpha}
\keyword{diversity}

