\name{moveme}
\alias{moveme}

\title{moveme}
\description{[TEXT]}
\usage{moveme=function (invec, movecommand)}
\arguments{
  \item{counts}{[TEXT]}
  \item{functional.map}{[TEXT]}
}
\details{[TEXT]}
\value{[TEXT]
}
\references{no references}
\author{dlemas}
\note{no further notes}

\seealso{no additional links}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
}

\keyword{ moveme }
