% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/alpha_diversity_calc.R
\name{alpha_diversity_calc}
\alias{alpha_diversity_calc}
\title{Calculate alpha diversity measures for each sample.}
\usage{
alpha_diversity_calc(otu_table, study_id, study_group)
}
\arguments{
\item{otu_table}{data frame with 16s counts}

\item{study_id}{vector that contains study id}

\item{study_group}{vector that contains syudy groups}
}
\description{
This function creates a data frame with alpha diversity estimates calculated for
each sample.
}
\examples{
# Simulated data
  head( )
  alpha_diversity_calc(otu.species.sort, meta.grp.sort$study_id, meta.grp.sort$bmi_group)
}
\keyword{alpha}
\keyword{diversity}

