% Generated by roxygen2 (4.0.1): do not edit by hand
\name{alpha_diversity_calc}
\alias{alpha_diversity_calc}
\title{Calculate alpha diversity measures for each sample.}
\usage{
alpha_diversity_calc(otu_table, meta_sample_name, meta_sort_group)
}
\arguments{
\item{otu_table}{data frame with 16s counts}

\item{meta_sample_name}{vector that contains sample names}

\item{meta_sort_group}{vector that contains sample names}
}
\description{
This function creates a data frame with alpha diversity estimates calculated for
each sample.
}
\examples{
# Simulated data
  head( )
  alpha_diversity_calc(otu.species.sort, meta.grp.sort$study_id, meta.grp.sort$bmi_group)
}
\keyword{alpha}
\keyword{diversity}

