% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/alpha_diversity_compare.R
\name{diversity_table_compare}
\alias{diversity_table_compare}
\title{Test for group differences in alpha diversity.}
\usage{
diversity_table_compare(meta_group, div_table)
}
\arguments{
\item{meta_group}{vector that contains sample groups}

\item{div_table}{data frame of alpha diversity measures for each sample}
}
\description{
This function tests for group differences in alpha diversity and returns a table
with alpha diversity mean/sd for all samples as well as group mean/sd.
}
\examples{
# Test
diversity_table_compare(table)
test=diversity_table_compare(meta.grp.sort$bmi_group,table); test
}
\keyword{alpha}
\keyword{diversity}

