\name{microbes-package}
\alias{microbes-package}
\alias{microbes}
\docType{package}
\title{This package analyzes data from next generation 16S sequencing}
\description{The functions in this package compute diversity and relative abundance comparisons between two groups}
\details{
\tabular{ll}{
Package: \tab microbes\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-07\cr
License: \tab What license is it under?\cr
}
}
\author{
dlemas and atomczik
Maintainer: dlemas
}
\references{refernces will be added here}
\keyword{ package }
\keyword{ microbiome }

\seealso{ put link here for tutorial}
\examples{}
