% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/cropping_functions.R
\name{prevalence_crop}
\alias{prevalence_crop}
\title{Crop data set by prevalence adn average relative abundance of bugs}
\usage{
prevalence_crop(df, x)
}
\arguments{
\item{df}{data frame with COG/KEGG counts}

\item{pathway.file}{array with pathways composed of COG/KEGG annotations}
}
\description{
This function will collapses COG/KEGG counts into metabolic pathways.
}
\examples{
# Simulated data
  head(df.wgs.counts)
  cogpathway=path.file
  df.wgs.pathway.counts=category2pathway(df.wgs.counts,cogpathway)
}
\keyword{pathway}

